// src/lib2/models/ApiResponseWrapperObject.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiResponseWrapperObject = {
    success?: boolean;
    message?: string;
    data?: Record<string, any>;
    timestamp?: string;
    traceId?: string;
    path?: string;
    code?: number;
};